Updated torque function so user can select landau or normal style
Using select case in fortran, the user can now select what style of torque they want to compute. The landau style
treats the torque like the elastic parameter is a landau parameter-- the more local energy, the higher the
temperature, and the elastic constant goes down. The upshot is that in regions where the spin field has high energy
the torque goes to almost zero-- this elevates the magnitude of the temperature flucuations, which will (ideally)
allow for defects to be spawned.